rename files for clarity and move adding of dataOutputAssociations to bpmn into it's class
